+++ changelog.md
@@ -1,9 +1,11 @@
-Changelog
+# Changelog
 
 All notable changes to this project will be documented in this file.
 This project adheres to semantic versioning.
 
 ## [unreleased]
+
+- option -f, --pattern-files now takes comma separated list of files
 
 ## [0.1.0] 2021-05-02